added 12 positions per second server to client websocket code

+++ server/ai.go
@@ -1,7 +1,7 @@
 package ai
 
 type Ai struct {
-	X, Y float
+	X, Y int
 }
 
 func (a *Ai) Foo() string {
